Fix workspace/symbol by removing protobuf generation.
SymbolData was intended to be persisted to disk, but we never made use
of it. The reason I removed the protobuf generation was in order to
add `SymbolData.info: SymbolInformation` to avoid untyped programming
again `flags: Int`. For example, workspace/symbol did not work because
there was a broken bitmask somewhere.
`SymbolInformation.{kind,properties}` are well
documented and should be used instead of the old unspecified flags.